requirements.txt
kagglehub
scikit-learn
numpy
tensorflow
matplotlib
opencv-python
sqlalchemy
python-dotenv

